requirements.txt
python-telegram-bot
openai
psycopg2-binary
python-dotenv
flask
duckduckgo-search
requests